// Public/src/models/RepairType.Model.ts
export interface IRepairTypeModel {
    Active: boolean
    AllocatedPrice: number
    Created: any
    CreatedBy: number
    DisplayInList: boolean
    DisplayName: string
    Enabled: any
    EnumName: string
    Id: number
    IsValid: boolean
    Modified: any
    ModifiedBy: number
    Name: string
    OfferPrice: number
    RowOrder: number
    ThumbnailPath: string
    ValidationErrors: {}
}